refactor(ai/core): rename stream language model call files
Rename stream model call implementation and test files to match streamLanguageModelCall naming, and add a changeset documenting the breaking type rename.

Made-with: Cursor
