+++ .changeset/rare-maps-invent.md
@@ -1,0 +1,7 @@
+---
+"ai": patch
+---
+
+refactor(ai/core): rename `ModelCallStreamPart` to `LanguageModelStreamPart` and align stream model call naming (`streamLanguageModelCall`, `experimental_streamLanguageModelCall`).
+
+This updates experimental low-level stream primitives to use "language model call" terminology consistently.
